[27] rename packages
